extract fakes to shared classes
