+++ app/models/crawler.rb
@@ -3,8 +3,8 @@ require 'public_suffix'
 require 'resque'
 
 class Crawler
-  def self.build(id, url)
-    new(url, GetURL, Resque)
+  def self.build(url)
+    new('a-uuid', url, GetURL, Resque)
   end
 
   def initialize(id, url, get, queue)
@@ -19,7 +19,7 @@ class Crawler
     [:internal_links, :external_links].each_with_index do |method, i|
       parser.send(method).each do |link|
         is_internal = i == 0
-        @queue.enqueue(LinkRepository, id, parser.domain, link, is_internal)
+        @queue.enqueue(LinkRepository, @id, parser.domain, link, is_internal)
       end
 
     end && true
