+++ test/helpers/fake_get.rb
@@ -1,0 +1,14 @@
+class FakeGet
+  attr_accessor :args, :calls
+  def initialize(html = nil)
+    @html = html
+  end
+  def do(url)
+    @args ||= []
+    @args << url
+    @calls ||= 0
+    @calls += 1
+
+    @html
+  end
+end
