+++ test/models/crawler_test.rb
@@ -1,5 +1,8 @@
 require 'minitest/autorun'
 
+require './test/helpers/fake_get'
+require './test/helpers/fake_queue'
+
 def html
   '
   <html>
@@ -15,31 +18,9 @@ class TestCrawler < Minitest::Test
   def setup
     url = 'http://bar.com'
 
-    # TODO: extract to a fakes lib as these are reused in other test
-    get = Struct.new(:foo) do
-      attr_accessor :args, :calls
-      def do(url)
-        @args ||= []
-        @args << url
-        @calls ||= 0
-        @calls += 1
-        html
-      end
-    end
-
-    queue = Struct.new(:foo) do
-      attr_accessor :args, :calls
-      def enqueue(*args)
-        @args ||= []
-        @args << args
-        @calls ||= 0
-        @calls += 1
-      end
-    end
-
-    @get = get.new(1)
-    @queue = queue.new(1)
-    @subj = Crawler.new(url, @get, @queue)
+    @get = FakeGet.new(html)
+    @queue = FakeQueue.new
+    @subj = Crawler.new('uuid', url, @get, @queue)
   end
 
   def test_perform
@@ -49,14 +30,14 @@ class TestCrawler < Minitest::Test
     assert_equal 6, @queue.calls, 'makes calls to enqueue'
 
     assert_equal(
-      [LinkRepository, 'bar.com', 'http://bar.com/qux1', true],
+      [LinkRepository, 'uuid', 'bar.com', 'http://bar.com/qux1', true],
       @queue.args[0],
       'adds internal links to persistence queue'
     )
 
     assert_equal(
-      [LinkRepository, 'bar.com', 'http://foo.com/bar', false],
-      @queue.args[4],
+      [LinkRepository, "uuid", "bar.com", "http://baz.com/bar", false],
+      @queue.args[5],
       'adds external links to persistence queue'
     )
   end
